+++ src/app/pin-map/pin-map.page.ts
@@ -234,10 +234,9 @@ export class PinMapPage {
     this.title.set(title);
     const timeRange = nowRange(this.db.getTimeZone());
     //const timeRange = nowRange(localTimeZone());
-    console.log('timeRange', timeRange);
     this.smallPins = true;
     const points = [];
-    const allEvents = await this.db.findEvents('', undefined, '', undefined, timeRange, false, false);
+    const allEvents = await this.db.findEvents('', undefined, '', undefined, timeRange, true, false);
     for (let event of allEvents) {
       const mapPoint = toMapPoint(
         event.location,